% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_data.R
\name{bootstrap_data}
\alias{bootstrap_data}
\title{Bootstrap data}
\usage{
bootstrap_data(data, speaker = NULL)
}
\arguments{
\item{data}{A data frame containing vowel formant data.}

\item{speaker}{A factor variable with subject IDs.}
}
\value{
A data frame.
}
\description{
Bootstrap data for use in cross-validation. Resampling with replacement is
done based on unique elements in the 'speaker' vector.
}
\examples{
\dontrun{
data(h95_data_no_NA)
boot <- bootstrap_data(h95_data_no_NA, speaker = h95_data_no_NA$speaker)
head(boot)
}
}

