% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cols}
\alias{cols}
\title{Colors for plotting}
\format{
A vector of colors for plotting
}
\usage{
cols
}
\description{
Colors for plotting
}
\examples{
\dontrun{
data(cols)
head(cols)
}
}
\keyword{datasets}

