% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svi_parallel.R
\name{combine_chunked_suffstats}
\alias{combine_chunked_suffstats}
\title{Combine Chunked Sufficient Statistics}
\usage{
combine_chunked_suffstats(chunk_results)
}
\arguments{
\item{chunk_results}{List of estep() results from parallel chunks}
}
\value{
Combined sufficient statistics list
}
\description{
Combines results from parallel E-step chunks into single sufficient
statistics structure. Used after parallel::mclapply() returns list of
estep results.
}
\keyword{internal}
