% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stm_svi.R
\name{compute_svi_defaults}
\alias{compute_svi_defaults}
\title{Compute Adaptive SVI Hyperparameters}
\usage{
compute_svi_defaults(N, K, V, has_prevalence)
}
\arguments{
\item{N}{Number of documents}

\item{K}{Number of topics}

\item{V}{Vocabulary size}

\item{has_prevalence}{Whether prevalence covariates are used}
}
\value{
List with adaptive hyperparameter values
}
\description{
Internal function to automatically select reasonable SVI hyperparameters
based on corpus characteristics (N, K, V) and covariate presence.
}
\keyword{internal}
