% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmvnorm_fast.R
\name{dmvnorm_fast}
\alias{dmvnorm_fast}
\title{Multivariate Normal Density}
\usage{
dmvnorm_fast(features, means, covariance, log = FALSE)
}
\arguments{
\item{features}{A matrix or vector of observations.}

\item{means}{A vector of means.}

\item{covariance}{A matrix of covariance.}

\item{log}{A boolean indicating whether to return the log density.}
}
\value{
A scalar, the density of the multivariate normal distribution.
}
\description{
Calculate the density of a multivariate normal distribution.
}
\examples{
\dontrun{
features <- matrix(c(500, 1500, 2500), nrow = 1)
means <- c(520, 1480, 2480)
covariance <- diag(c(10000, 25000, 40000))
dmvnorm_fast(features, means, covariance, log = TRUE)
}
}

