% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heldout.R
\name{eval.heldout}
\alias{eval.heldout}
\alias{make.heldout}
\title{Heldout Likelihood by Document Completion}
\usage{
eval.heldout(model, missing)
}
\arguments{
\item{model}{an stm model}

\item{missing}{a missing object created by make.heldout}
}
\description{
Tools for making and evaluating heldout datasets.
}
\details{
These functions are used to create and evaluate heldout likelihood using the
document completion method.  The basic idea is to hold out some fraction of
the words in a set of documents, train the model and use the document-level
latent variables to evaluate the probability of the heldout portion. See the
example for the basic workflow.
}
\examples{

\dontrun{

prep <- prepDocuments(poliblog5k.docs, poliblog5k.voc, 
                      poliblog5k.meta,subsample=500,
                      lower.thresh=20,upper.thresh=200)
heldout <- make.heldout(prep$documents, prep$vocab)
documents <- heldout$documents
vocab <- heldout$vocab
meta <- out$meta

stm1<- stm(documents, vocab, 5, 
           prevalence =~ rating+ s(day), 
           init.type="Random",
           data=meta, max.em.its=5)
eval.heldout(stm1, heldout$missing)
}
}

