% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral.R
\name{fastAnchor}
\alias{fastAnchor}
\title{Find Anchor Words}
\usage{
fastAnchor(Qbar, K, verbose = TRUE)
}
\arguments{
\item{K}{The number of desired anchors}

\item{verbose}{If TRUE prints a dot to the screen after each anchor}

\item{Q}{The gram matrix}
}
\description{
Take a gram matrix Q and returns K anchors.
}
\details{
Does not use any randomness. Notes that anchors are such that you can request
more than you want but its not particularly easy to simply start in the middle
of a process.  One possibility for allowing this would be to return Qbar but
I'm not going to worry about that right now.
}

