% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_posterior.R
\name{find_posterior}
\alias{find_posterior}
\title{Find the posterior distribution}
\usage{
find_posterior(likelihoods, priors, type = "BSTM")
}
\arguments{
\item{likelihoods}{A matrix of likelihoods with columns for the mean, standard deviation, and density.}

\item{priors}{A matrix of priors with columns for the mean, standard deviation, and density.}

\item{type}{A string indicating the type of method to use for estimating the posterior.}
}
\value{
A matrix of posteriors with columns for the mean, standard deviation, density, probability, and rounded probability.
}
\description{
This function calculates the posterior distribution for a given set of
likelihoods and priors. It accepts a matrix of likelihoods and a matrix of
priors and returns a matrix of posteriors. The function also accepts a string
indicating the type of method to use for estimating the posterior. The
function returns a matrix of posteriors with columns for the mean, standard
deviation, density, probability, and rounded probability.
}
\examples{
\dontrun{
likelihoods <- matrix(c(7.1, 0.12, -1.0, 7.3, 0.11, -0.8), ncol = 3, byrow = TRUE)
colnames(likelihoods) <- c("likelihood_mu", "likelihood_sd", "likelihood_density")
priors <- matrix(c(7.2, 0.13, -0.5, 7.2, 0.13, -0.5), ncol = 3, byrow = TRUE)
colnames(priors) <- c("prior_mu", "prior_sd", "prior_density")
find_posterior(likelihoods, priors, type = "BSTM")
}
}

