% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findTopic.R
\name{findTopic}
\alias{findTopic}
\title{Find topics that contain user specified words.}
\usage{
findTopic(
  x,
  list,
  n = 20,
  type = c("prob", "frex", "lift", "score"),
  verbose = TRUE
)
}
\arguments{
\item{x}{The STM model object to be searched. May also be the output from
sageLabels.}

\item{list}{Character vector containing words to be searched.}

\item{n}{Number of words to consider}

\item{type}{Type of words to be searched.}

\item{verbose}{A logical indicating whether details should be printed to the
screen.}
}
\description{
Find topics that contain user specified words.
}
\examples{

lab <- sageLabels(gadarianFit, n=5)
findTopic(lab, c("poor", "immigr", "peopl"))
findTopic(gadarianFit, c("poor", "immigr", "peopl"))

}
\seealso{
\code{\link{findThoughts}}
}
