% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitNewDocuments.R
\name{fitNewDocuments}
\alias{fitNewDocuments}
\title{Fit New Documents}
\usage{
fitNewDocuments(model, documents, data = NULL, betaIndex = NULL,
  prevalencePrior = c("Average", "Covariate", "None"),
  contentPrior = c("Average", "Covariate"), returnPosterior = FALSE,
  returnPriors = FALSE)
}
\arguments{
\item{model}{the originally fit STM object.}

\item{documents}{these documents must be in the stm format and be numbered in the same way
as the documents in the original model with the same dimension of vocabulary!}

\item{data}{the metadata for the prevalence prior which goes with the unseen documents}

\item{betaIndex}{an integer vector which indicates which level of the content covariate is used
for each unseen document.}

\item{prevalencePrior}{three options described in detail below.  Defaults to "Average" when
\code{data} is not provided and "Covariate" when it is.}

\item{contentPrior}{two options described in detail below. Defaults to "Average" when 
\code{betaIndex} is not provided and "Covariate" when it is.}

\item{returnPosterior}{the function always returns the posterior mode of theta
(document-topic proportions),  If set to TRUE this will return the full variational
posterior.  Note that this will return a dense K-by-K matrix for every document
which can be very memory intensive if you are processing a lot of documents.}

\item{returnPriors}{the function always returns the options that were set for the prior
(either by the user or chosen internally by the defaults).  In the case of content covariates
using the covariate prior this will be a set of indices to the original beta matrix so as
not to make the object too large.}
}
\description{
A function for predicting thetas for an unseen document based on the previously fit model.
}
\details{
Due to the existence of the metadata in the model, this isn't as simple as in models
without side information such as Latent Dirichlet Allocation. There are four scenarios:
models without covariate information, models with prevalence covariates only, models with
content covariates only and models with both.  When there is not covariate information the
choice is essentially whether or not to use prior information.

For each case we offer three choices (and may offer more in the future):
\describe{
\item{"None"}{No prior is used.  In the prevalence case this means that the model simply
maximizes the likelihood of seeing the words given the word-topic distribution.  This will
in general produce more sharply peaked topic distributions than the prior. This can be used
even without the covariates.  This is not an option for topical content covariate models.  If
you do not observe the topical content covariate, use the "Average" option.} 
\item{"Average"}{We use a prior that is based on the average over the documents in the training
set.  This does not require the unseen documents to observe the covariates.  In a model that originally
had covariates we need to adjust our estimate of the variance-covariance matrix sigma to accomodate that
we no longer have the covariate information.  So we recalculate the variance based on what it would have
been if we didn't have any covariates.  This helps avoid an edge case where the covariates are extremely
influential and we don't want that strength applied to the new covariate-less setting.  In the case of 
content covariates this essentially use the \code{\link{sageLabels}} approach to create a
marginalized distribution over words for each topic.
}
\item{"Covariate"}{We use the same covariate driven prior that existed in the original model.
This requires that the test covariates be observed for all previously unseen documents.}
}

If you fit a document that was used during training with the options to replicate the initial
\code{\link{stm}} model fit you will not necessarily get exactly the same result.  \code{\link{stm}}
updates the topic-word distributions last so they may shifted since the document-topic proportions
were updated.  If the original model converged, they should be very close.
}
\examples{

}

