% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_posterior.R
\name{get_posterior}
\alias{get_posterior}
\title{Get posterior probabilities from a list of BSTM results}
\usage{
get_posterior(results_list)
}
\arguments{
\item{results_list}{A list of BSTM results}
}
\value{
A matrix of posterior probabilities with columns for each category
and rows for each token.
}
\description{
This function takes a list of BSTM results and returns a matrix of posterior
probabilities. The function returns a matrix with columns for each category
and rows for each token.
}
\examples{
\dontrun{
data(h95_data_no_NA)
template <- create_template(h95_data_no_NA[, 4:6], h95_data_no_NA$vowel)
res <- BSTM(h95_data_no_NA[1:5, 4:6], h95_data_no_NA$f0[1:5], template)
post <- get_posterior(res)
head(post)
}
}

