% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateEffect.R
\name{glance.estimateEffect}
\alias{glance.estimateEffect}
\title{Glance at an object produced by `stm::estimateEffect`}
\usage{
\method{glance}{estimateEffect}(x, ...)
}
\arguments{
\item{x}{An object returned by the `stm::estimateEffect` function}

\item{...}{extra arguments (not used)}
}
\value{
A data.frame with one column per property of the estimateEffect object
}
\description{
Glance at an object produced by `stm::estimateEffect`
}
