% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral.R
\name{gram}
\alias{gram}
\title{Compute the gram matrix}
\usage{
gram(mat)
}
\arguments{
\item{mat}{A Matrix sparse Document by Term matrix}
}
\description{
Take a Matrix object and compute a gram matrix
}
\details{
Due to numerical error in floating points you can occasionally get
very small negative values.  Thus we check if the minimum is under 0 
and if true, assign elements less than 0 to zero exactly.  This is mostly
so we don't risk numerical instability later by introducing negative numbers of 
any sort.
}

