% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{h95_classifications}
\alias{h95_classifications}
\title{Hillenbrand et al. (1995) Classifications}
\format{
A data frame with 1168 rows and 12 columns:
\describe{
  \item{\code{ae}}{Number of classifications into ae.}
  \item{\code{ah}}{Number of classifications into ah.}
  \item{\code{aw}}{Number of classifications into aw.}
  \item{\code{eh}}{Number of classifications into eh.}
  \item{\code{ei}}{Number of classifications into ei.}
  \item{\code{er}}{Number of classifications into er.}
  \item{\code{ih}}{Number of classifications into ih.}
  \item{\code{iy}}{Number of classifications into iy.}
  \item{\code{oa}}{Number of classifications into oa.}
  \item{\code{oo}}{Number of classifications into oo.}
  \item{\code{uh}}{Number of classifications into uh.}
  \item{\code{uw}}{Number of classifications into uw.}
}
}
\source{
\url{https://example.com/dataset_source}
}
\usage{
h95_classifications
}
\description{
Classification data data for the H95 dataset from 20 listeners presented the vowels at random.
}
\examples{
\dontrun{
data(h95_classifications)
str(h95_classifications)
head(h95_classifications)
}
}
\keyword{datasets}

