% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lobanov_normalize.R
\name{lobanov_normalize}
\alias{lobanov_normalize}
\title{Lobanov normalization}
\usage{
lobanov_normalize(formants, speakers, vowels, warning = FALSE)
}
\arguments{
\item{formants}{A matrix of formant data.}

\item{speakers}{A vector of speaker IDs.}

\item{vowels}{A vector of vowel IDs.}

\item{warning}{A boolean indicating whether to display a warning message.}
}
\value{
A data frame of normalized formant data.
}
\description{
This function normalizes formant data using Lobanov normalization.
This method is not recommended for most linguistic research.
For more information on this, please see:
}
\details{
Barreda, S. (2021). Perceptual validation of vowel normalization methods
for variationist research. Language Variation and Change, 33(1), 27-53.
}
\examples{
\dontrun{
data(h95_data_no_NA)
formants <- as.matrix(h95_data_no_NA[, 4:6])
out <- lobanov_normalize(formants, h95_data_no_NA$speaker, h95_data_no_NA$vowel)
head(out)
}
}

