% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_methods.R
\name{method6}
\alias{method6}
\title{Method 6 psi estimation}
\usage{
method6(
  ffs,
  f0,
  template,
  PSI_prior_mean = 7.233,
  PSI_prior_sd = 0.1284,
  f0_hat_sd = 0.1327,
  f0_hat_intercept = -10.32,
  f0_hat_slope = 2.145,
  vowel_priors = vowel_priors,
  correctOUflow = TRUE,
  type = "BSTM",
  ...
)
}
\arguments{
\item{ffs}{A vector of formant frequencies}

\item{f0}{A scalar of fundamental frequency}

\item{template}{A template object containing category means and covariance information}

\item{PSI_prior_mean}{A scalar of the prior mean for psi}

\item{PSI_prior_sd}{A scalar of the prior standard deviation for psi}

\item{f0_hat_sd}{A scalar of the standard deviation for f0 given psi}

\item{f0_hat_intercept}{A scalar of the intercept for f0 given psi}

\item{f0_hat_slope}{A scalar of the slope for f0 given psi}

\item{vowel_priors}{A vector of prior probabilities for each category}

\item{correctOUflow}{A boolean indicating whether to correct for underflow}

\item{type}{A string indicating the type of method to use for estimating psi}

\item{...}{additional parameters}
}
\value{
A scalar of the estimated psi value, and optionally the log density at that location.
}
\description{
This function estimates the psi value using the method 6 from Nearey and Assmann (2007).
It also returns the log density of the estimated psi value.
}
\examples{
\dontrun{
data(h95_data_no_NA)
template <- create_template(h95_data_no_NA[, 4:6], h95_data_no_NA$vowel)
method6(as.numeric(h95_data_no_NA[1, 4:6]), f0 = h95_data_no_NA$f0[1], template = template)
}
}

