% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_likelihood.R
\name{model_likelihood}
\alias{model_likelihood}
\title{Model likelihood from posterior probabilities}
\usage{
model_likelihood(
  classifications,
  posterior_probabilities,
  correctOUflow = TRUE
)
}
\arguments{
\item{classifications}{A vector of classifications}

\item{posterior_probabilities}{A vector of posterior probabilities.}

\item{correctOUflow}{A boolean indicating whether to correct for underflow.}
}
\value{
A scalar, the likelihood of the model given the classifications and posterior probabilities.
}
\description{
Calculates an overall model likelihood score from observed classifications and
their corresponding posterior probabilities.
}
\examples{
\dontrun{
classifications <- c("a", "i", "a", "u")
posterior_probabilities <- c(0.8, 0.7, 0.6, 0.9)
model_likelihood(classifications, posterior_probabilities)
}
}





