% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyTopics.R
\name{paretosingle}
\alias{paretosingle}
\title{Function to choose the pareto dominant model}
\usage{
paretosingle(z)
}
\arguments{
\item{z}{An stm model object from \code{selectModel()}.}
}
\description{
Function to choose the pareto dominant model, or randomly choose
among model candidates that are not weakly dominanted by other
models utility function for manyTopics.
}
