% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_posterior.R
\name{plot_posterior}
\alias{plot_posterior}
\title{Plot posterior density from STM output}
\usage{
plot_posterior(df, psi_range = c(6.9, 7.6), n_steps = 1000)
}
\arguments{
\item{df}{A dataframe of STM output}

\item{psi_range}{A vector of the range of psi values to plot}

\item{n_steps}{A scalar of the number of steps to take in the psi range}
}
\description{
Plots posterior density curves from a single \code{STM_output} data frame,
including vowel-specific contributions across a psi grid.
}
\examples{
\dontrun{
data(h95_data_no_NA)
template <- create_template(h95_data_no_NA[, 4:6], h95_data_no_NA$vowel)
out <- BSTM(as.numeric(h95_data_no_NA[1, 4:6]), h95_data_no_NA$f0[1], template)
plot_posterior(out$df)
}
}





