% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSTM.R
\name{PSTM}
\alias{PSTM}
\title{Probabilistic Sliding Template Model}
\usage{
PSTM(
  ffs,
  f0,
  template,
  method = method6,
  vowel_priors = NULL,
  correctOUflow = TRUE,
  ...
)
}
\arguments{
\item{ffs}{A vector, matrix, or dataframe of formant frequencies. Each row
is a different observation, each column is a formant measurement.}

\item{f0}{A vector of fundamental frequencies}

\item{template}{A template object}

\item{method}{A function that that estimates the 'best' psi given some information.}

\item{vowel_priors}{A vector of prior probabilities for each category}

\item{correctOUflow}{A boolean indicating whether to correct for underflow
and overflow using the correctOUflow() function.}

\item{...}{Additional arguments are passed to internal call of estimation method.}
}
\value{
In the case of a single token, a dataframe of psi estimates and
posterior probabilites. If several tokens (i.e. rows) are passed, output is
a list of dataframes.
}
\description{
Estimate psi from individual tokens using an update to the algorithm
suggested by Nearey and Assmann (2007). Returns posterior probabilities for
each category based on the estimate of psi, the template, and token
acoustic properties.
}
\examples{
\dontrun{
data(h95_data_no_NA)
template <- create_template(h95_data_no_NA[, 4:6], h95_data_no_NA$vowel)
ffs <- as.numeric(h95_data_no_NA[1, 4:6])
f0 <- h95_data_no_NA$f0[1]
out <- PSTM(ffs, f0, template)
print(out)
}
}

