% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral.R
\name{recoverL2}
\alias{recoverL2}
\title{RecoverL2}
\usage{
recoverL2(Qbar, anchor, wprob, verbose = TRUE, ...)
}
\arguments{
\item{Qbar}{The row-normalized gram matrix}

\item{anchor}{A vector of indices for rows of Q containing anchors}

\item{wprob}{The empirical word probabilities used to renorm the mixture weights.}

\item{verbose}{If TRUE prints information as it progresses.}

\item{...}{Optional arguments that will be passed to the exponentiated gradient algorithm.}
}
\value{
\item{A}{A matrix of dimension K by V.  This is acturally the transpose of A in Arora et al. and the matrix we call beta.}
}
\description{
Recover the topic-word parameters from a set of anchor words using the RecoverL2
procedure of Arora et. al.
}
\details{
Using the exponentiated algorithm and an L2 loss identify the optimal convex 
combination of the anchor words which can reconstruct each additional word in the 
matrix.  Transform and return as a beta matrix.
}

