% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STMfunctions.R
\name{rmvnorm}
\alias{rmvnorm}
\title{Draw from a Multivariate Normal}
\usage{
rmvnorm(n, mu, Sigma, chol.Sigma = chol(Sigma))
}
\arguments{
\item{n}{number of draws}

\item{mu}{the K-dimensional mean}

\item{Sigma}{the K by K dimensional positive definite covariance matrix}

\item{chol.Sigma}{the cholesky decomposition of the Sigma matrix.}
}
\description{
A basic function for doing multivariate normal simulations
via the cholesky decomposition of the covariance matrix. Function
is based on one by Peter Hoff.
}
\details{
This is a pretty standard multivariate normal generator. It could
almost certainly be faster if we ported it over to \pkg{RcppArmadillo}
but it isn't used a ton at the moment.
}
\keyword{internal}
