% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_posterior.R
\name{scale_posterior}
\alias{scale_posterior}
\title{Scale posterior probabilities to sum to 1}
\usage{
scale_posterior(posterior, type = "BSTM")
}
\arguments{
\item{posterior}{A matrix of posterior probabilities
with columns for each category and rows for each token.}

\item{type}{A string indicating how the posterior should be calculated,
either "BSTM" or "PSTM".}
}
\value{
A matrix of posterior probabilities with an additional column
representing the scaled posterior probabilities.
}
\description{
This function scales the posterior probabilities to sum to 1.
It accepts a matrix of posterior probabilities and add columns to the matrix
representing posteriors, and modifying posterior densities as required.
}
\examples{
\dontrun{
posterior <- matrix(c(7.1, 0.10, -1.2, 7.3, 0.12, -0.8), ncol = 3, byrow = TRUE)
scale_posterior(posterior, type = "BSTM")
}
}

