% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STM_output.R
\name{STM_output}
\alias{STM_output}
\alias{print.STM_output}
\alias{[.STM_output}
\alias{plot.STM_output}
\title{Create an object of class "STM_output"}
\usage{
STM_output(df, stimulus, template, parameters, vowel_priors)

\method{print}{STM_output}(x, digits = 3, ...)

\method{[}{STM_output}(x, i, j)

\method{plot}{STM_output}(x, y, ...)
}
\arguments{
\item{df}{the dataframe}

\item{stimulus}{the stimulus}

\item{template}{the template}

\item{parameters}{the parameters}

\item{vowel_priors}{the vowel priors}

\item{x}{An object of class "STM_output".}

\item{digits}{Number of digits to print.}

\item{...}{Additional arguments passed to other methods.}

\item{i}{Row index.}

\item{j}{Column index.}

\item{y}{unused parameter}
}
\value{
An object of class "STM_output".

A subset of the data frame in the STM_output object.
}
\description{
This function creates an object of the S3 class "STM_output", the output of the XSTM functions.
}
\examples{
\dontrun{
data(h95_data_no_NA)
template <- create_template(h95_data_no_NA[, 4:6], h95_data_no_NA$vowel)
out <- BSTM(as.numeric(h95_data_no_NA[1, 4:6]), h95_data_no_NA$f0[1], template)
class(out)
}
}

