% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STM_output_list.R
\name{STM_output_list}
\alias{STM_output_list}
\alias{print.STM_output_list}
\alias{$.STM_output_list}
\alias{[[.STM_output_list}
\alias{length.STM_output_list}
\alias{sapply.STM_output_list}
\title{Create an object of class "STM_output_list"}
\usage{
STM_output_list(output_list)

\method{print}{STM_output_list}(x, ...)

\method{$}{STM_output_list}(x, name)

\method{[[}{STM_output_list}(x, i)

\method{length}{STM_output_list}(x)

\method{sapply}{STM_output_list}(X, FUN, ..., simplify = TRUE, USE.NAMES = TRUE)
}
\arguments{
\item{output_list}{A list of STM_output objects.}

\item{x}{An object of class "STM_output_list".}

\item{...}{Additional arguments passed to other methods.}

\item{name}{Element name.}

\item{i}{index.}

\item{X}{An object of class "STM_output_list".}

\item{FUN}{The function to be applied.}

\item{simplify}{A logical indicating whether to simplify the result.}

\item{USE.NAMES}{A logical indicating whether to use names.}
}
\value{
An object of class "STM_output_list".

A subset of the data frame in the STM_output_list object.
}
\description{
Creates an S3 \code{STM_output_list} object from multiple token-level
\code{STM_output} results and exposes convenience accessors for summaries.
}
\examples{
\dontrun{
data(h95_data_no_NA)
template <- create_template(h95_data_no_NA[, 4:6], h95_data_no_NA$vowel)
res <- BSTM(h95_data_no_NA[1:5, 4:6], h95_data_no_NA$f0[1:5], template)
length(res)
head(res$winners)
}

}

