% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STM_posteriors.R
\name{STM_posteriors}
\alias{STM_posteriors}
\alias{print.STM_posteriors}
\alias{[.STM_posteriors}
\alias{apply.STM_posteriors}
\title{Create an object of class "STM_posteriors"}
\usage{
STM_posteriors(mat)

\method{print}{STM_posteriors}(x, digits = 3, ...)

\method{[}{STM_posteriors}(x, i, j)

\method{apply}{STM_posteriors}(X, MARGIN, FUN, ...)
}
\arguments{
\item{mat}{the matrix of posteriors}

\item{x}{An object of class "STM_posteriors".}

\item{digits}{Number of digits to print.}

\item{...}{Additional arguments passed to other methods.}

\item{i}{Row index.}

\item{j}{Column index.}

\item{X}{An object of class "STM_posteriors".}

\item{MARGIN}{A vector giving the subscripts which the function will be applied over.}

\item{FUN}{The function to be applied.}
}
\value{
An object of class "STM_posteriors".

A subset of the data frame in the STM_posteriors object.
}
\description{
This function creates an object of the S3 class "STM_posteriors", the output of the XSTM functions.
}
\examples{
\dontrun{
mat <- matrix(runif(12), nrow = 3)
mat <- mat / rowSums(mat)
post <- STM_posteriors(mat)
summary(post)
}
}

