% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateEffect.R
\name{tidy.estimateEffect}
\alias{tidy.estimateEffect}
\title{Tidy an object produced by `stm::estimateEffect`}
\usage{
\method{tidy}{estimateEffect}(x, ...)
}
\arguments{
\item{x}{An object returned by the `stm::estimateEffect` function}

\item{...}{extra arguments (not used)}
}
\value{
A data.frame with columns for topic numbers, coefficient names,
standard errors, t statistics, and p-values
}
\description{
Tidy an object produced by `stm::estimateEffect`
}
