% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STMfunctions.R
\name{unpack.glmnet}
\alias{unpack.glmnet}
\title{Unpack a \pkg{glmnet} object}
\usage{
unpack.glmnet(mod, ic.k)
}
\arguments{
\item{mod}{the glmnet model}

\item{ic.k}{the information criterion value.  AIC is \code{ic.k=2} and BIC would be \code{ic.k=log n}}
}
\value{
A list
\item{coef}{a matrix of coefficients}
\item{intercept}{the intercepts}
}
\description{
A function to quickly unpack a \pkg{glmnet} model object and calculate an
optimal model from the regularization path.
}
\details{
This is a small utility we wrote to deal with the slow methods dispatch for S4
classes.  The more straightforward option is the \code{coef()} method for \pkg{glmnet}
objects but when trying to make thousands of calls a second, that can be very slow
}
\keyword{internal}
