% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_gaussians.R
\name{combine_gaussians}
\alias{combine_gaussians}
\title{Function to combine two Gaussian distributions}
\usage{
combine_gaussians(mu_1, sd_1, mu_2, sd_2, d_1 = NULL, d_2 = NULL)
}
\arguments{
\item{mu_1}{The mean of the first Gaussian distribution}

\item{sd_1}{The standard deviation of the first Gaussian distribution}

\item{mu_2}{The mean of the second Gaussian distribution}

\item{sd_2}{The standard deviation of the second Gaussian distribution}

\item{d_1}{The peak density of the first gaussian}

\item{d_2}{The peak density of the second gaussian}
}
\value{
The mean and standard deviation of the combined Gaussian distribution.
}
\description{
This function combines two Gaussian distributions by taking the mean and
standard deviation of the two distributions and returning a new Gaussian
distribution that is the combination of the two.
}
\examples{
# TBD
}
