% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctOUflow_internal.R
\name{correctOUflow_internal}
\alias{correctOUflow_internal}
\title{correctOUflow_internal}
\usage{
correctOUflow_internal(posterior_probabilities)
}
\arguments{
\item{posterior_probabilities}{A vector of posterior probabilities}
}
\value{
A vector of posterior probabilities
}
\description{
Correct for underflow and overflow in posterior probabilities
}
