% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_template.R
\name{create_template}
\alias{create_template}
\alias{print.STM_template}
\alias{plot.STM_template}
\title{Create a Template}
\usage{
create_template(features, classes, shared_covar = FALSE)

\method{print}{STM_template}(x, ...)

\method{plot}{STM_template}(x, ...)
}
\arguments{
\item{features}{A matrix or data frame of acoustic information. Each row is a
different observation, each column is a formant measurement, f0, duration,
or any other token characteristic.}

\item{classes}{A vector of class labels for each observation.}

\item{shared_covar}{A boolean indicating whether to use a shared covariance
matrix for all classes.}

\item{x}{A template object.}

\item{...}{Additional arguments.}
}
\value{
A template object.
}
\description{
Create a dialectal template for use with the (X)STM family of functions.
}
\examples{
# TBD
}
