% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_ellipse.R
\name{draw_ellipse}
\alias{draw_ellipse}
\title{Draw an ellipse}
\usage{
draw_ellipse(means, covariance, n_points = 100, show = TRUE, ...)
}
\arguments{
\item{means}{A vector of length 2 with the x and y coordinates of the center of the ellipse}

\item{covariance}{A 2x2 matrix with the covariance matrix of the ellipse}

\item{n_points}{The number of points to use to draw the ellipse}

\item{show}{A boolean indicating whether to plot the ellipse}

\item{...}{Additional arguments to pass to the lines function}
}
\description{
Draw an ellipse
}
\examples{
# TBD
}
