% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_likelihood.R
\name{estimate_likelihood}
\alias{estimate_likelihood}
\title{Estimate the likelihood of a given set of ffs}
\usage{
estimate_likelihood(ffs, means, covariance, precision = solve(covariance))
}
\arguments{
\item{ffs}{A vector of formant frequencies}

\item{means}{A vector of category means}

\item{covariance}{A matrix of category covariance}

\item{precision}{A matrix of category precision}
}
\value{
A list of the likelihood mean, likelihood standard deviation, and likelihood density.
}
\description{
This function estimates the likelihood of a given set of formant frequencies
given a set of category means and covariance matrices. It returns the mean,
standard deviation, and density of the likelihood.
}
\examples{
# TBD
}
