% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_posterior.R
\name{find_posterior}
\alias{find_posterior}
\title{Find the posterior distribution}
\usage{
find_posterior(likelihoods, priors, type = "BSTM")
}
\arguments{
\item{likelihoods}{A matrix of likelihoods with columns for the mean, standard deviation, and density.}

\item{priors}{A matrix of priors with columns for the mean, standard deviation, and density.}

\item{type}{A string indicating the type of method to use for estimating the posterior.}
}
\value{
A matrix of posteriors with columns for the mean, standard deviation, density, probability, and rounded probability.
}
\description{
This function calculates the posterior distribution for a given set of
likelihoods and priors. It accepts a matrix of likelihoods and a matrix of
priors and returns a matrix of posteriors. The function also accepts a string
indicating the type of method to use for estimating the posterior. The
function returns a matrix of posteriors with columns for the mean, standard
deviation, density, probability, and rounded probability.
}
\examples{
# TBD
}
