% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_winners.R
\name{get_winners}
\alias{get_winners}
\title{Get the winners of the BSTM model}
\usage{
get_winners(results_list)
}
\arguments{
\item{results_list}{A list of results from a BSTM or PSTM model.}
}
\value{
A data frame with columns for the psi and vowel.
}
\description{
This function returns the winning psi and vowel for each token in the BSTM model.
It accepts a list of results and returns a data frame with columns for the psi
and vowel.
}
\examples{
# TBD
}
