% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inpute_NA.R
\name{impute_NA}
\alias{impute_NA}
\title{impute_NA}
\usage{
impute_NA(formants, speakers, vowels, sim_error = FALSE)
}
\arguments{
\item{formants}{A matrix of formant frequencies}

\item{speakers}{A vector of speaker IDs}

\item{vowels}{A vector of vowel IDs}

\item{sim_error}{A logical indicating whether to simulate error.}
}
\value{
What your function returns.
}
\description{
Imputes missing formant values with or without random error.
}
\details{
Predicts log(formant)~speaker+vowel and imputes missing values.
If sim_error is TRUE, imputes with random error by simulating normal
data with a mean of zero and a standard deviation equal to the residual
error of the regression model.
}
\examples{
# Example usage:
# TBD
}
