% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_numerical.R
\name{integrate_numerical}
\alias{integrate_numerical}
\title{Numerical integration}
\usage{
integrate_numerical(x, y)
}
\arguments{
\item{x}{A vector of x values}

\item{y}{A vector of y values}
}
\value{
A scalar of the integral of the function.
}
\description{
This function calculates the integral of a function using the trapezoidal
rule. The function iaccepts a set of x and y values and returns the
integral of the function.
}
\examples{
# TBD
}
