% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_methods.R
\name{method2}
\alias{method2}
\title{Method 2 psi estimation}
\usage{
method2(
  ffs,
  f0 = NA,
  template,
  vowel_priors = NULL,
  correctOUflow = TRUE,
  type = "BSTM",
  ...
)
}
\arguments{
\item{ffs}{A vector of formant frequencies}

\item{f0}{A scalar of fundamental frequency}

\item{template}{A template object containing category means and covariance information}

\item{vowel_priors}{A vector of prior probabilities for each category}

\item{correctOUflow}{A boolean indicating whether to correct for underflow}

\item{type}{A string indicating the type of method to use for estimating psi}

\item{...}{additional parameters}
}
\value{
A scalar of the estimated psi value, and optionally the log density at that location.
}
\description{
This function estimates the psi value using the method 3 from Nearey and Assmann (2007).
It also returns the log density of the estimated psi value.
}
\examples{
#examples
# TBD
}
