% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_likelihood.R
\name{model_likelihood}
\alias{model_likelihood}
\title{model_likelihood}
\usage{
model_likelihood(
  classifications,
  posterior_probabilities,
  correctOUflow = TRUE
)
}
\arguments{
\item{classifications}{A vector of classifications}

\item{posterior_probabilities}{A vector of posterior probabilities.}

\item{correctOUflow}{A boolean indicating whether to correct for underflow.}
}
\value{
A scalar, the likelihood of the model given the classifications and posterior probabilities.
}
\description{
Calculate the likelihood of a model given classifications and posterior probabilities
}
\examples{
# Example usage:
# TBD
}
