% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Normalize formant data}
\usage{
normalize(formants, speakers, vowels)
}
\arguments{
\item{formants}{A matrix of formant frequencies}

\item{speakers}{A vector of speaker IDs}

\item{vowels}{A vector of vowel IDs}
}
\value{
A matrix of normalized formant data
}
\description{
This function normalizes formant data using the least-squares method
described in Barreda and Nearey (2018). This method is robust in the presence
of to missing and unbalanced data. Log-mean normalized data and speaker
psi estimates are returned.
}
\examples{
#examples
# TBD
}
