% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_posterior.R
\name{plot_posterior}
\alias{plot_posterior}
\title{Plot STM modelling results
Plot the posterior distribution of the formants and the fundamental frequency
given a template and a set of formant frequencies.}
\usage{
plot_posterior(df, psi_range = c(6.9, 7.6), n_steps = 1000)
}
\arguments{
\item{df}{A dataframe of STM output}

\item{psi_range}{A vector of the range of psi values to plot}

\item{n_steps}{A scalar of the number of steps to take in the psi range}
}
\description{
Plot STM modelling results
Plot the posterior distribution of the formants and the fundamental frequency
given a template and a set of formant frequencies.
}
\examples{
# TBD
}
