% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s.R
\name{s}
\alias{s}
\title{Make a B-spline Basis Function}
\usage{
s(x, df, ...)
}
\arguments{
\item{x}{The predictor value.}

\item{df}{Degrees of freedom.  Defaults to the minimum of 10 or one minus
the number of unique values in x.}

\item{\dots}{Arguments passed to the \code{\link[splines]{bs}} function.}
}
\value{
A predictor matrix of the basis functions.
}
\description{
This is a simple wrapper around the \code{\link[splines]{bs}} function in
the splines package.  It will default to a spline with 10 degrees of
freedom.
}
\details{
This is a simple wrapper written as users may find it easier to simply type
\code{s} rather than selecting parameters for a spline. We also include
\code{predict} and \code{makepredictcall} generic functions for the class
so it will work in settings where \code{\link{predict}} is called.
}
\seealso{
\code{\link[splines]{bs}} \code{\link[splines]{ns}}
}

