% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STM.R
\name{STM}
\alias{STM}
\title{Sliding Template Model}
\usage{
STM(formants, template, vowel_priors = NULL, correctOUflow = TRUE)
}
\arguments{
\item{formants}{A vector of formant frequencies, or a matrix or data frame of formant vectors. Each row
is a different observation, each column is a formant measurement.}

\item{template}{A template object}

\item{vowel_priors}{A vector of prior probabilities for each vowel category.}

\item{correctOUflow}{A boolean indicating whether to correct for underflow
and overflow using the correctOUflow() function.}
}
\value{
A dataframe of posterior probabilities, one row for each token.
a list of dataframes.
}
\description{
Returns posterior probabilities for each category based on the template and
token acoustic properties.
}
\examples{
# TBD
}
