% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STM_output_list.R
\name{STM_output_list}
\alias{STM_output_list}
\alias{print.STM_output_list}
\alias{[[.STM_output_list}
\alias{sapply.STM_output_list}
\title{Create an object of class "STM_output_list"}
\usage{
STM_output_list(output_list)

\method{print}{STM_output_list}(x, ...)

\method{[[}{STM_output_list}(x, i)

\method{sapply}{STM_output_list}(X, FUN, ..., simplify = TRUE, USE.NAMES = TRUE)
}
\arguments{
\item{output_list}{A list of STM_output objects.}

\item{x}{An object of class "STM_output_list".}

\item{...}{Additional arguments passed to other methods.}

\item{i}{index.}

\item{X}{An object of class "STM_output_list".}

\item{FUN}{The function to be applied.}

\item{simplify}{A logical indicating whether to simplify the result.}

\item{USE.NAMES}{A logical indicating whether to use names.}
}
\value{
An object of class "STM_output".

A subset of the data frame in the STM_output_list object.
}
\description{
This function creates an object of the S3 class "STM_output_list", the a list of outputs
of the XSTM functions.
}
\examples{
# TBD
# TBD

}
