% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.STM.R
\name{summary.STM}
\alias{summary.STM}
\alias{print.STM}
\title{Summary Function for the STM objects}
\usage{
\method{summary}{STM}(object, ...)
}
\arguments{
\item{object}{An STM object.}

\item{\dots}{Additional arguments affecting the summary}
}
\description{
Function to report on the contents of STM objects
}
\details{
Summary prints a short statement about the model and then runs
\code{\link{labelTopics}}.
}
